Fix: Separate mapping for sft
