+++ src/autotrain/trainers/clm/__main__.py
@@ -17,27 +17,27 @@ def train(config):
     # if isinstance(config, dict):
     #     config = LLMTrainingParams(**config)
     if config.trainer == "default":
-        from autotrain.trainers.clm.train_clm_default import train as train_default
+        from autotrain.trainers.clm.default.train_clm_default import train as train_default
 
         train_default(config)
 
     elif config.trainer == "sft":
-        from autotrain.trainers.clm.train_clm_sft import train as train_sft
+        from autotrain.trainers.clm.sft.train_clm_sft import train as train_sft
 
         train_sft(config)
 
     elif config.trainer == "reward":
-        from autotrain.trainers.clm.train_clm_reward import train as train_reward
+        from autotrain.trainers.clm.reward.train_clm_reward import train as train_reward
 
         train_reward(config)
 
     elif config.trainer == "dpo":
-        from autotrain.trainers.clm.train_clm_dpo import train as train_dpo
+        from autotrain.trainers.clm.dpo.train_clm_dpo import train as train_dpo
 
         train_dpo(config)
 
     elif config.trainer == "orpo":
-        from autotrain.trainers.clm.train_clm_orpo import train as train_orpo
+        from autotrain.trainers.clm.orpo.train_clm_orpo import train as train_orpo
 
         train_orpo(config)
